Add Services section to Home.razor with dynamic content
Introduced a new "Services" section to the `Home.razor` file, displaying a list of services offered by the company in a card format. Each card includes a title and description of the service. The services are dynamically generated using a `foreach` loop that iterates over a list of `Service` objects.

- Added a heading "What would you like to get done?" for the new section.
- Defined a private list of `Service` objects within the `@code` block.
- Created a `Service` class with `Title` and `Description` properties.
- Used a `foreach` loop to generate HTML for each service card.