Add MaxLength attribute to ImageName in UploadModel
The ImageName property in the UploadModel class now includes a MaxLength attribute with a maximum length of 45 characters. An error message "Length should be less than 46 letters" is provided for validation. This change enforces a constraint on the length of the image name to maintain data integrity and user input validation.